Moved MatrixIO to model package (where it should be!)
